% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate.R
\name{accumulate}
\alias{accumulate}
\alias{accumulate.harp_spatial_fcst}
\title{Compute accumulation between forecast lead times}
\usage{
accumulate(.fcst, accum_time, ...)

\method{accumulate}{harp_spatial_fcst}(.fcst, accum_time, .fcst_name = "", ...)
}
\arguments{
\item{.fcst}{A harp_fcst list or a harp_spatial_fcst data frame.}

\item{accum_time}{The time over which to compute the accumulation. Must be in
the same units as the lead_time column.}

\item{...}{Further arguments passed to or from methods.}

\item{.fcst_name}{Not to be set by the user, but used for consistency between
classes.}
}
\value{
An object of the same class as .fcst with the accumulations computed.
Rows for lead times for which it is not possible to compute the
accumulation are dropped.
}
\description{
For data with equally spaced lead times, the difference between, or
accumulation of gridded fields between lead times is computed. The time over
which to compute the accumulation must be set and is calculated from the lead
times. If lead times are not equally spaced an attempt will be made to
compute all accumulations for the desired accumulation time from the
available lead times.
}
\details{
This function would typically be used compute accumulations between lead
times for model outputs that are accumulated from the model start time, such
as precipitation and fluxes. However, \code{accumulate} could also be used
to compute the change in any value between lead times, such a change in
temperature or change in wind speed for example.
}
