% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_forecast.R
\name{accumulate_forecast}
\alias{accumulate_forecast}
\title{Accumulate a forecast over a given time period.}
\usage{
accumulate_forecast(.fcst, accumulation_time, accumulation_unit)
}
\arguments{
\item{.fcst}{A data frame of class \code{harp_point_forecast}.}

\item{accumulation_time}{The time over which the accumulation is to be
calculated. Must be in the same time units as the \code{leadtime} column.}
}
\value{
A data frame of class \code{harp_point_forecast} with the
  \code{forecast} column containing the accumulated forecast.
}
\description{
Given an object of class \code{harp_point_forecast},
\code{accumulate_forecast} will return the same object, but with the forecast
column having the amount accumulated over the specified
\code{accumulation_time}. This is done through subtraction of the forecast at
lead time T - \code{accumulation_time} from the forecast at lead time T.
\code{accumulation_time} must therefore be in the same units as the
\code{leadtime} column.
}

