% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_harp_parameter.R
\name{as_harp_parameter}
\alias{as_harp_parameter}
\title{Create a harp parameter}
\usage{
as_harp_parameter(
  fullname,
  basename,
  level = 0,
  level_type = c("unknown", "height", "msl", "surf", "pressure", "model"),
  accum = 0,
  accum_unit = NULL
)
}
\arguments{
\item{fullname}{The full name of the parameter}

\item{basename}{The basename of the parameter. This is typically the fullname
excluding accumulation time and vertical level infotmation from the
parameter name and forms the basis of the parameter name that is searched
for in files. For example, for reading parameters from grib files the
basename should be exactly the same as the grib shortName.}

\item{level}{The vertical level of the parameter. For grib files, this is
ignored if the basename begins with a number, or mx or mn followed by a number.}

\item{level_type}{The type of vertical level. Can be one of "unknown",
"height", "msl", "surf", "pressure", or "model".}

\item{accum}{The accumulation time in seconds}

\item{accum_unit}{The unit of accumulation to be used when searching files.}
}
\value{
A harp paramater
}
\description{
harp includes a number of built in parameters, but you may wish to read other
parameters. To do this you need to create a harp parameter.
}
\examples{
# maximum 2m temperature in last 6 hours
as_harp_parameter("mx2t6")
# potential vorticity at 500 hPa
as_harp_parameter("pv500", "pv", level = 500, level_type = "pressure")
}
