% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_fcst.R
\name{bind_fcst}
\alias{bind_fcst}
\title{Bind deterministic forecasts into a single data frame}
\usage{
bind_fcst(.fcst)
}
\arguments{
\item{.fcst}{A harp_fcst list}
}
\value{
A data frame
}
\description{
Given a harp_fcst list, \code{bind_fcst} binds the data frames in the
harp_fcst list into a single data frame for easy plotting. For ensemble
forecasts, the member data are gathered into a single column.
}
