% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_utils.R
\name{create_table}
\alias{create_table}
\title{Create a new table in an SQLite data base}
\usage{
create_table(db, name, data, primary = NULL, show_query = FALSE)
}
\arguments{
\item{db}{A database connection}

\item{name}{A name for the table. If it already exists, nothing happens.}

\item{primary}{Primary keys}

\item{a}{data.frame. Only column names and type are used. Alternatively, it
can also be a named vector of date types.}
}
\description{
Create a new table in an SQLite data base
}
