% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_mean_and_var.R
\name{ens_mean_and_var}
\alias{ens_mean_and_var}
\title{Compute the ensemble mean and variance}
\usage{
ens_mean_and_var(
  .fcst,
  mean_name = "ens_mean",
  var_name = "ens_var",
  sd_name = "ens_spread",
  var_drop_member = NULL
)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} object, or a data frame containing columns
with "_mbr" as part of the column names.}

\item{mean_name}{The output column name for the ensemble mean}

\item{var_name}{The output column name for the ensemble variance}

\item{sd_name}{The output column name for the ensemble spread (standard
deviation)}

\item{var_drop_member}{Which members to drop for the calculation of the
ensemble variance and standard deviation. For harp_fcst objects, this can
be a numeric scalar - in which case it is recycled for all forecast models;
a list or numeric vector of the same length as the harp_fcst object, or a
named list with the names corresponding to names in the harp_fcst object.}

\item{...}{}
}
\value{
A \code{harp_fcst} object with columns ens_mean and ens_var added to
the forecast tables.
}
\description{
The ensemble mean and variance are computed and added as columns to tables in
a \code{harp_fcst} object.
}
