% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_stats.R
\name{ens_stats}
\alias{ens_stats}
\title{Get ensemble statistics}
\usage{
ens_stats(
  .fcst,
  mean = TRUE,
  spread = TRUE,
  var = FALSE,
  min = FALSE,
  max = FALSE,
  prob_thresh = NULL,
  prob_inequality = `>=`,
  nbh_radius = 0
)
}
\arguments{
\item{.fcst}{A harp_fcst list or a harp_point_fcst or harp_spatial_fcst data
frame.}

\item{mean}{Logical. Set to TRUE (the default) to compute the ensemble mean.}

\item{spread}{Logical. Set to TRUE (the default) to compute the ensemble
spread.}

\item{var}{Logical. Set to TRUE to compute the ensemble variance.}

\item{min}{Logical. Set to TRUE to compute the ensemble minimum.}

\item{max}{Logical. Set to TRUE to compute the ensemble maximum.}

\item{prob_thresh}{A numeric threshold for which to compute the probability.
If a length 2 vector is passed, the probability for between those two
values (inclusive) will be computed.}

\item{prob_inequality}{The inequality to use in the calculation for the
probability. Must be in back ticks. Can be any one of `>=` (the default),
`>`, `<=`, `<`, `==`, or `!=`.}

\item{nbh_radius}{For spatial data the radius of neighbourhoods for which to
compute the probability in grid lengths. The neigbourhood will be a square
with sides equal 2 * nbh_radius.}
}
\value{
An object of the same class as .fcst with new columns in the data
  frame(s) for the computed values.
}
\description{
Get ensemble statistics
}
