% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fa.R, R/read_fatar.R
\name{fa_opts}
\alias{fa_opts}
\alias{fatar_opts}
\title{Set options for FA decoding}
\usage{
fa_opts(
  meta = TRUE,
  fa_type = "arome",
  fa_vector = TRUE,
  rotate_wind = TRUE,
  ...
)

fatar_opts(ldt_template = "+\%04i$", ...)
}
\arguments{
\item{meta}{If TRUE, the time and grid details are also decoded. This is slower.}

\item{fa_type}{The kind of model file: "arome", "alaro", "surfex"... Mainly important for precipitation fields.}

\item{fa_vector}{TRUE if the wind variable (speed, direction) must be calculated from components U & V}

\item{rotate_wind}{TRUE means wind U,V (along axes of the grid) should be rotated to actual N.}

\item{...}{Any non-standard options that don't have default values.}

\item{ldt_template}{a character string that describes (part of) the file name used for finding a lead time within a tar archive.}
}
\value{
Returns a list of options. Either the defaults or any modification.
}
\description{
Set options for FA decoding
}
