% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{fcparquet_opts}
\alias{fcparquet_opts}
\alias{obsparquet_opts}
\title{Options for writing parquet datasets.}
\usage{
fcparquet_opts(
  path,
  partitioning = c("fcst_hour", "fcst_year", "fcst_month", "fcst_day"),
  dir_mode = "0750",
  ...
)

obsparquet_opts(
  path,
  partitioning = c("valid_year", "valid_month", "valid_day"),
  dir_mode = "0750",
  ...
)
}
\arguments{
\item{path}{The path to which to write the dataset.}

\item{partitioning}{How the data are to be partitioned in the directory
structure. This should not be changed.}

\item{dir_mode}{The numeric permission mode of the directories to be written.
The default is "0750".}

\item{...}{Not used.}
}
\value{
A named list.
}
\description{
Parquet datasets can be used as point data storage for harp as an
alternative to SQLite. A parquet dataset is a multi-file dataset made up
of parquet files stored in directories that are partitioned based on the
data. This makes it effeicient to query the data and tends to result in
faster performance than SQLite and less use of storage space. Internally, the
\href{https://arrow.apache.org/docs/r/index.html}{arrow} package is used to
read from and write to parquet datasets. \code{fcparquet_opts()} and
\code{obsparquet_opts()} generate options for writing point forecast using
\code{\link{read_forecast()}} and \code{\link{read_obs()}} respectively.
}
