% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_params.R
\name{file_parameter_definitions}
\alias{file_parameter_definitions}
\alias{new_grib_param}
\alias{new_netcdf_param}
\alias{new_fa_param}
\alias{new_v_param}
\alias{new_obsoul_param}
\title{Make a parameter definition list for a file format}
\usage{
new_grib_param(name, level_type = get("grib_level_types"), level = NULL)

new_netcdf_param(name, suffix = get("nc_level_suffixes"))

new_fa_param(name, units, name_len = 16)

new_v_param(name, harp_param, units, type = c("SYNOP", "TEMP"))

new_obsoul_param(name, harp_param, units)
}
\arguments{
\item{name}{The name of the parameter in the file. In the case of grib files,
this should be the grib shortName and in the case of obsoul files, should be
the number used to define what the variable is. In addition, can be expressed
as a character vector for grib files in order to try different shortNames
until the paramatet is found. Should be a named list if a function is to be
applied with the names matching the argument names in the function.}

\item{level_type}{The grib levelType. Can be a character vector in order to
try different levelTypes until the variable is found in the file. The
default is a named list using the in built grib_level_types to define the
names for diffferent vertical coordinate systems.}

\item{level}{The level number. Set to -999 for all levels.}

\item{suffix}{The suffix to add to \code{name} for different vertical
coordinate systems.}

\item{units}{The units of the parameter in vfld / vobs files.}

\item{name_len}{The number of characters that \code{name} should have in the
FA files. This is 16 by default and adds blank spaces to the end of
\code{name} to make it the required length.}

\item{harp_param}{The name of the parameter to be used in outputs from read
functions.}

\item{type}{The type of the parmater - either "SYNOP" or "TEMP"}
}
\description{
These functions are to be used together with \link{add_param_def} and
\link{modify_param_def} to add a parameter definition entry for a particular
file format.
}
