% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fa_param_info.R
\name{get_fa_param_info}
\alias{get_fa_param_info}
\title{FA parameter names}
\usage{
get_fa_param_info(
  param,
  fa_type = "arome",
  fa_vector = TRUE,
  rotate_wind = TRUE
)
}
\arguments{
\item{param}{Parameter name}

\item{fa_type}{could be 'alaro' or 'arome'. These models store e.g.
precipitation under a different name.}

\item{fa_vector}{A logical value. TRUE means that e.g. wind speed will be
calculated from wind vector components, not searched as a separate field.}

\item{rotate_wind}{If TRUE, wind fields (U, V, direction)
are rotated from model axes to earth axes.}
}
\value{
A 16 character string, or in rare cases a vector of several strings
  denoting the components (e.g. total precipitation may be the sum of up to 4
  fields).
}
\description{
a harp-style parameter is translated to the field name in
ALARO/AROME/HARMONIE model output in FA format
}
