% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netcdf_param_MET.R
\name{get_netcdf_param_MET}
\alias{get_netcdf_param_MET}
\title{Get MET Norway NetCDF parameter name from a HARP parameter name}
\usage{
get_netcdf_param_MET(param)
}
\arguments{
\item{param}{HARP parameter name.}
}
\value{
The parameter name in a MET Norway NetCDF file
}
\description{
Get MET Norway NetCDF parameter name from a HARP parameter name
}
\examples{
get_netcdf_param_MET("T2m")
get_netcdf_param_MET("PMSL")
get_netcdf_param_MET("S10m")
get_netcdf_param_MET("Pcp")
}

