% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Get File Name Template}
\usage{
get_template(template)
}
\arguments{
\item{template}{The file type to generate the template for. Can be
"harmoneps_grib", "harmeoneps_grib_fp", "harmoneps_grib_sfx", "meps_met",
"meps_smhi", "meps_fmi", "arome_arctic", "harmonie_grib",
"harmonie_grib_fp", "harmone_grib_sfx", "vfld", "vobs", or "fctable". If
anything else is passed, it is returned unmodified. The function is case
insensitive.}
}
\value{
A template that can be interpreted by \code{glue_data} from the
  \code{glue} package.
}
\description{
Get File Name Template
}
\examples{
get_template("harmoneps_grib")
get_template("harmonie_grib_fp")

}

