% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grid.R
\name{guess_format}
\alias{guess_format}
\title{Try to guess the binary format of a gridded data file.}
\usage{
guess_format(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
A character string with the format, or NA.
}
\description{
Try to guess the binary format of a gridded data file.
}
