% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harp_params}
\alias{harp_params}
\title{Parameter definitions}
\format{
A list of 60 parameters that have entries for different file formats
includeing grib, netcdf, fa, v (cfld/vobs) and obsoul.
}
\usage{
harp_params
}
\description{
A list of parameters with name definitions for different file formats.
}
\keyword{datasets}
