% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpIO-package.R
\docType{package}
\name{harpIO-package}
\alias{harpIO}
\alias{harpIO-package}
\title{harpIO: IO functions and data wrangling for harp}
\description{
Functions for reading and writing data in the harp framework. Readable formats include grib, netCDF, FA, vfld, vobs, SQLite FCTABLE files and writeable formats are SQLite FCTABLE files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/harphub/harpIO}
  \item Report bugs at \url{https://github.com/harphub/harpIO/issues}
}

}
\author{
\strong{Maintainer}: Andrew Singleton \email{andrewts@met.no}

Authors:
\itemize{
  \item Alex Deckmyn \email{alex.deckmyn@meteo.be}
}

}
\keyword{internal}
