% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_opts.R
\name{hdf5_opts}
\alias{hdf5_opts}
\title{Set Options for Reading HDF5 files}
\usage{
hdf5_opts(
  data_path = NULL,
  odim = TRUE,
  meta = TRUE,
  invert_data = TRUE,
  iflag = NULL,
  pcp_quant = "ACRR",
  pinfo = FALSE
)
}
\arguments{
\item{data_path}{This option is not respected! Notionaly this is the
path to the data inside the hdf5 file, e.g. obtained from h5dump -n, but
by default all datasets in the hdf5 are read and are then matched to the
parameter specified by 'parameter' in \code{read_hdf5}.}

\item{odim}{Logical. This option is not functional! Notionally this indicates
if the file follows the ODIM standard, but is not respected at the moment.}

\item{meta}{Logical. Get metavar such as domain information from the hdf5
file. If set to TRUE then a geofield will be returned, if FALSE just the
raw data will be returned. Generally should be TRUE.}

\item{invert_data}{Logical. Should the data be inverted i.e. transposed and
put upside-down?}

\item{iflag}{An institute flag to indicate specific decoding of hdf5 files.
The only currently recognised option is 'knmi'. If NULL then the ODIM
standard is assumed.}

\item{pcp_quant}{The name used for precipitation in the hdf5 file. Defaults
to ACCR i.e. the ODIM standard.}

\item{pinfo}{Logical. Print some info from \code{read_hdf5} for debugging.}
}
\value{
A list of options for reading hdf5 files.
}
\description{
When reading a hdf5 file in harp, the system needs to know some information
about the file format. These options are then used in \code{read_hdf5}, which
generally assumes that the hdf5 file uses the ODIM standard.
}
\details{
If your hdf5 file differs from the ODIM standard, you can specify specific
decoding through the \code{iflag} (if it has been added to \code{read_hdf5}).
The only recognised option for \code{iflag} is currently 'knmi'.
The variable name used to recognise precipitation in the hdf5 file defaults
to 'ACCR'. This can be changed by using \code{pcp_quant}.

The default options can be seen by running \code{hdf5_opts()}.
}
\examples{
hdf5_opts()
hdf5_opts(invert_data = FALSE, iflag = "knmi", pcp_quant = "PRECIP_[MM]")
}
