% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_to_pressure.R
\name{hybrid_to_pressure}
\alias{hybrid_to_pressure}
\title{Title}
\usage{
hybrid_to_pressure(
  .df,
  level_definitions,
  psfc_col = psfc,
  a_col = ap,
  b_col = b
)
}
\arguments{
\item{psfc_col}{}
}
\description{
Title
}
