% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_interpolation.R
\name{initialise_interpolation}
\alias{initialise_interpolation}
\title{Initialise domain, interpolation, land/sea mask etc. for a model}
\usage{
initialise_interpolation(clim_file = NULL, file_format = "fa",
  domain = NULL, stations = NULL, method = "closest",
  use_mask = FALSE, correct_t2m = FALSE, drop_NA = TRUE, ...)
}
\arguments{
\item{domain}{A \code{geodomain}, only needed if it can not be retrieved from clim file.}

\item{stations}{The locations for which to initialise interpolation}

\item{method}{Interpolation method ("bilin", "nn"...)}

\item{use_mask}{Should L/S mask be applied when initialising the interpolation?}

\item{drop_NA}{Should weights/stations that are NA be removed?}

\item{...}{Arguments for interpolation method. Currently not used.}

\item{climfile}{The path of a "clim" file containing topography & land/sea mask}

\item{ftype}{File type of the clim file (fa, grib ...).}
}
\value{
A list with various initialisations. It is also assigned to the global environment.
}
\description{
Initialise domain, interpolation, land/sea mask etc. for a model
}
