% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_psfc_to_xsection.R
\name{join_psfc_to_xsection}
\alias{join_psfc_to_xsection}
\title{Title}
\usage{
join_psfc_to_xsection(
  xsection,
  psfc_param = "ps",
  fcst_model = "psfc",
  file_path = getwd(),
  file_template = "harmonie_grib",
  file_format = NULL,
  file_format_opts = list(),
  show_progress = FALSE
)
}
\arguments{
\item{show_progress}{}
}
\description{
Title
}
