% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_opts.R
\name{modify_opts}
\alias{modify_opts}
\title{Modify an options list}
\usage{
modify_opts(opts, ..., add_options = FALSE)
}
\arguments{
\item{opts}{The options list to be modified.}

\item{...}{Expressions that give new values for options in \code{opts}.}

\item{add_options}{If \code{add_options = TRUE}, options that are not already
found in \code{opts} will be added to \code{opts}. The default is FALSE and
will give a warning for each option not found in \code{opts}}
}
\description{
harpIO includes a number of functions for creating lists of options for, for
example, file formats and transformations. \code{modify_opts} allows you to
change the values for given options in an already generated options list.
}
\details{
\code{modify_opts} is designed to be used in a pipeline so that an options
list can easily be modified. It is particularly useful for modifying options
set via, for example, an options set in \code{\link{netcdf_opts}}.
}
\examples{
# Default value for z_var is "height1" with
# options_set = "met_norway_eps"
netcdf_opts(options_set = "met_norway_eps")

# Change z_var to "height0", keeping all other
# options from options_set = "met_norway_eps"
netcdf_opts(options_set = "met_norway_eps") \%>\%
  modify_opts(z_var = "height2")
}
