% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msub.R
\name{msub}
\alias{msub}
\title{Multiple pattern matching and replacement}
\usage{
msub(x, pattern, replacement, regex = FALSE, ...)
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A character vector of patterns to replace.}

\item{replacement}{A character vector of the same length as pattern with the
corresponding replacements.}

\item{regex}{A logical to denote whether \code{pattern} is a regular
expression (TRUE) or a string to match exactly (FALSE). The default is
FALSE so strings in \code{pattern} are matched exactly.}

\item{...}{Other arguments to \link[base]{gsub}}
}
\value{
A character vector of the same length as x
}
\description{
\code{msub} calls \link[base]{gsub} for each pattern / replacement pair.
Where a pattern is not found, nothing is replaced for that pattern /
replacement pair.
}
\details{
When regex is set to FALSE strings in \code{pattern} are matched exactly.
This means that only full strings are matched and not substrings.
}
\examples{
msub(letters[1:10], c("a", "c", "e"), c("A", "C", "E"))
msub(c("a", "b", "ac", "ad"), c("a", "ac"), c("A", "AC"))
\dontrun{
# "ac" is replaced with "Ac" due to substring matching
msub(c("a", "b", "ac", "ad"), c("a", "ac"), c("A", "AC"), regex = TRUE)
}
}
