% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_parse.R
\name{param_parse}
\alias{param_parse}
\title{parse the parameter name and decide on level, accumulation...}
\usage{
param_parse(param)
}
\arguments{
\item{param}{Parameter name}
}
\value{
A list with \code{fullname}, \code{basename}, \code{level_type},
  \code{level}, \code{accum}
}
\description{
Parses a harp-style parameter name and returns accumulation time, level (if
any) and the base name of the parameter
}
\examples{
param_parse("AccPcp1h")
param_parse("z500")
param_parse("s100m")
}

