% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_harp_parameter.R
\name{parse_harp_parameter}
\alias{parse_harp_parameter}
\title{Parse a harp paramter.}
\usage{
parse_harp_parameter(param)
}
\arguments{
\item{param}{The parameter name.}
}
\value{
A list with the following elements: \cr
- fullname: the original input name (==param) \cr
- basename: the name without any accumulation info \cr
- shortName: as found by grib_api (and/or Rgrib2) \cr
- levelType: either numeric (100,105 etc) or S,P,H or NULL \cr
- level: NULL if not needed (e.g. if shortName is enough) or numerical \cr
- accum: the accumulation period (0 or higher) \cr
- FAname : the corresponding name in FA format files \cr
}
\description{
\code{parse_harp_parameter} returns vertical level information, accumulation time and parameter names
for grib and FA reading.
}
\examples{
parse_harp_parameter("T2m")
parse_harp_parameter("AccPcp12h")
parse_harp_parameter("Z500")
parse_harp_parameter("U100m")
parse_harp_parameter("Q61s")

}

