% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_v_parameter.R
\name{parse_v_parameter}
\alias{parse_v_parameter}
\title{Parameters for vobs and vfld files}
\usage{
parse_v_parameter(v_parameter, param_type, param_defs = get("harp_params"))
}
\arguments{
\item{v_parameter}{vobs or vfld parameter}

\item{param_type}{type of parameter - "SYNOP" or "TEMP"}
}
\value{
A list with the harp paramater name and the units of the parameter.
}
\description{
Internal function to convert parameter names from vobs or vfld files into
harp format parameters. Use parse_v_parameter_synop for SYNOP parameters and
parse_v_parameter_temp for TEMP parameters.
}
