% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_v_parameter_synop.R
\name{parse_v_parameter_synop}
\alias{parse_v_parameter_synop}
\title{Parse a synop parameter from a vobs or vfld convention to harp convention}
\usage{
parse_v_parameter_synop(v_parameter_synop)
}
\arguments{
\item{v_parameter_synop}{vobs or vfld parameter}
}
\value{
harp paramater name
}
\description{
Parse a synop parameter from a vobs or vfld convention to harp convention
}
\examples{
parse_v_parameter_synop("TT")
parse_v_parameter_synop("RH")
parse_v_parameter_synop("FI")
parse_v_parameter_synop("PE")

}

