% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_v_parameter_temp.R
\name{parse_v_parameter_temp}
\alias{parse_v_parameter_temp}
\title{Parse a temp parameter from a vobs or vfld convention to harp convention}
\usage{
parse_v_parameter_temp(v_parameter_temp)
}
\arguments{
\item{v_parameter_synop}{vobs or vfld parameter}
}
\value{
harp paramater name
}
\description{
Parse a temp parameter from a vobs or vfld convention to harp convention
}
\examples{
parse_v_parameter_synop("TT")
parse_v_parameter_synop("RH")
parse_v_parameter_synop("FI")
parse_v_parameter_synop("PP")

}

