% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_fa.R
\name{prm_fa}
\alias{prm_fa}
\title{FA parameter names}
\usage{
prm_fa(param, type = "alaro")
}
\arguments{
\item{param}{Parameter name}

\item{type}{could be 'alaro' or 'arome'. These models store e.g. precipitation under a
different name.}
}
\value{
A 16 character string, or in rare cases a vector of several strings denoting the
  components (e.g. total precipitation may be the sum of up to 4 fields).
}
\description{
a harp-style parameter is translated to the field name in
ALARO/AROME/HARMONIE model output in FA format
}

