% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eps_interpolate.R
\name{read_eps_interpolate}
\alias{read_eps_interpolate}
\title{Title}
\usage{
read_eps_interpolate(start_date, end_date, eps_model, parameter, file_path,
  lead_time = seq(0, 48, 3), members_in = seq(0, 9), by = "6h",
  file_format = "vfld", file_template = "vfld", sqlite_path = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

