% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fa.R
\name{read_fa}
\alias{read_fa}
\title{Read a field from a FA file}
\usage{
read_fa(infile, parameter, meta = TRUE, ...)
}
\arguments{
\item{parameter}{The parameter to read. Standard HARP names are used.}

\item{meta}{If TRUE, also read all meta data (domain, time properties).}

\item{...}{Arguments for \code{Rgrib2::Gdec}}

\item{filename}{The FA file name.}
}
\value{
A geofield object with 2d array and projection information
}
\description{
Read a field from a FA file
}
\examples{
model_geofield <- read_fa(file_name, "t2m")
model_geofield <- read_fa(file_name, "t500")
model_geofield <- read_fa(file_name, "topo")
}

