% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fa.R
\name{read_fa_interpolate}
\alias{read_fa_interpolate}
\title{Read FA files & interpolate}
\usage{
read_fa_interpolate(file_name, parameter, lead_time = NA_real_,
  members = NA_character_, vertical_coordinate = NA_character_,
  init = list(), method = "closest", use_mask = FALSE,
  fa_type = "arome", fa_vector = TRUE, ...)
}
\arguments{
\item{file_name}{Name of a tar archive containing FA files}

\item{parameter}{The parameter(s) to be decoded.}

\item{lead_time}{The lead time(s) to be extracted. May be a vector!}

\item{members}{Mostly ignored, but could be added as a (constant) column to output.
If present it must be a single string value (FA files do not contain multiple ensemble members)}

\item{vertical_coordinate}{Not used. Only there for API reasons.}

\item{init}{Interpolation weights (and domain information).}

\item{method}{Interpolation method (only necessary if the weights are not yet initialised)}

\item{use_mask}{If TRUE, use land/sea mask in interpolation}

\item{fa_type}{For some fields (e.g. precipitation) arome and alaro
use different names, so we should specify.}

\item{fa_vector}{If true, wind speed will be calculated from U and V components.}

\item{...}{Ignored and simply passed to read_fatar}

\item{...}{Extra arguments for read_fa[tar]}
}
\value{
A tibble with interpolated forecasts for the stations list
}
\description{
Read FA files & interpolate
}
