% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fatar.R
\name{read_fatar}
\alias{read_fatar}
\title{Read a field from an FA file in a tar archive}
\usage{
read_fatar(file_name, parameter, lead_time = 0, levels = NULL,
  members = NULL, fa_type = "arome", fa_vector = TRUE,
  lt_unit = "h", ...)
}
\arguments{
\item{file_name}{A tar archive of FA files.}

\item{parameter}{The parameter to read. Standard HARP names are used,
but full FA field names will also work. If NULL, only domain information is read.}

\item{lead_time}{Expressed in hours.}

\item{...}{Arguments for \code{read_fa}}
}
\value{
A 2d geofield object (2d array with projection information).
        If the parameter is not found, the geofield will have value NA.
}
\description{
Read a field from an FA file in a tar archive
}
\examples{
model_geofield <- read_fatar(file_name, "t2m", lead_time=0)
model_geofield <- read_fa(file_name, "t500", lead_time=6)
}
