% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fatar.R
\name{read_fatar_interpolate}
\alias{read_fatar_interpolate}
\title{Read fa-tar files and interpolate to a set of locations}
\usage{
read_fatar_interpolate(file_name, parameter, lead_time, members = NULL,
  init = list(), method = "closest", use_mask = FALSE,
  fa_type = "arome", fa_vector = TRUE, ...)
}
\arguments{
\item{file_name}{Name of a tar archive containing FA files}

\item{parameter}{The parameter(s) to be decoded.}

\item{lead_time}{The lead time(s) to be extracted. May be a vector!}

\item{members}{Mostly ignored, but could be added as a (constant) column to output.
If present it must be a single string value (FA files do not contain multiple ensemble members)}

\item{init}{Interpolation weights (and domain information).}

\item{method}{Interpolation method (only necessary if the weights are not yet initialised)}

\item{use_mask}{If TRUE, use land/sea mask in interpolation}

\item{...}{Ignored and simply passed to read_fatar}
}
\value{
a list of two tibbles. One with interpolated forecasts for the stations list,
        and one with parameter units.
}
\description{
Read fa-tar files and interpolate to a set of locations
}
