% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grib.R
\name{read_grib}
\alias{read_grib}
\title{Read a field from a grib file}
\usage{
read_grib(file_name, parameter, meta = TRUE, ...)
}
\arguments{
\item{file_name}{The grib file name.}

\item{parameter}{The parameter to read. Standard HARP names are used.}

\item{meta}{If TRUE, also read all meta data (domain, time properties).}

\item{...}{Arguments for \code{Rgrib2::Gdec}}
}
\value{
A geofield object with 2d array and projection information
}
\description{
Read a field from a grib file
}
