% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grib.R
\name{read_grib_interpolate}
\alias{read_grib_interpolate}
\title{Read a field from a grib file & interpolate}
\usage{
read_grib_interpolate(file_name, parameter, lead_time = NA_real_,
  members = NA_character_, vertical_coordinate = NA_character_,
  init = list(), method = "closest", use_mask = FALSE, ...)
}
\arguments{
\item{file_name}{The grib file name.}

\item{parameter}{The parameter to read. Standard HARP names are used.}

\item{lead_time}{lead time}

\item{members}{ens members}

\item{vertical_coordinate}{Not yet used.}

\item{init}{Initialisation for interpolation. A list that contains
station locations and (possibly) pre-calculated interpolation weights etc.}

\item{method}{Interpolation method (only necessary if the weights are not yet initialised)}

\item{use_mask}{If TRUE, use land/sea mask in interpolation}

\item{...}{Arguments for \code{Rgrib2::Gdec}}

\item{meta}{If TRUE, also read all meta data (domain, time properties).}
}
\value{
A tibble
}
\description{
Read a field from a grib file & interpolate
}
