% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grid.R
\name{read_grid}
\alias{read_grid}
\alias{read_grid_interpolate}
\title{Read gridded data in various formats.}
\usage{
read_grid(filename, parameter, file_format = NULL, data_frame = FALSE, ...)

read_grid_interpolate(filename, parameter, file_format = NULL, ...)
}
\arguments{
\item{filename}{A character string with the (full) file name.}

\item{parameter}{The parameter to be read.}

\item{file_format}{The file format. Possible values include grib, netcdf, FA,
hdf5... Whatever the value is, it is suposed to correspond to a function
"read_XXX" that can deal with the format. If not specified, the format can
often be guessed correctly from file extension or the first few bytes.}

\item{data_frame}{Logical. For compatibility with current version of
harpSpatial, make sure only a geofield is returned rather than a data
frame when set to FALSE.}

\item{...}{All arguments passed to the specified reader function.}
}
\value{
A geofield or (possibly) a plain  matrix.
}
\description{
Read gridded data in various formats.
}
\examples{
if (requireNamespace("Rgrib2", quietly = TRUE) & requireNamespace("harpData", quietly = TRUE)) {
  read_grid(
    system.file("grib/HARMUK20171015T12Z+003.grib", package = "harpData"),
    parameter = "T2m"
  )
  read_grid(
    system.file("grib/HARMUK20171015T12Z+003.grib", package = "harpData"),
    parameter = "RH2m"
  )
  read_grid(
    system.file("grib/HARMUK20171015T12Z+003.grib", package = "harpData"),
    parameter = "tcc"
  )
}
}
