% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf5.R
\name{read_hdf5}
\alias{read_hdf5}
\title{Read/decode hdf5 files, mainly ODIM style}
\usage{
read_hdf5(filename, data = "dataset1/data1/data", meta = TRUE)
}
\description{
Read/decode hdf5 files, mainly ODIM style
}

