% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_netcdf.R
\name{read_netcdf}
\alias{read_netcdf}
\title{Read a 2d field from NetCDF file.}
\usage{
read_netcdf(filename, parameter, member, lead_time, level = NULL)
}
\arguments{
\item{filename}{NetCDF file name to read from.}

\item{parameter}{Parameter to read - may be a parameter in the netcdf file
or a standard HARP parameter name.}

\item{member}{The member to read from the file.}

\item{lead_time}{The lead time to read from the file.}

\item{level}{The pressure level for the data}
}
\value{
A 2-d array.
}
\description{
\code{read_netcdf} returns a 2d field from a NetCDF file for a given
member and lead time. This function has only been tested on MEPS data.
The levels argument will currently only get data on pressure levels.
}
\examples{
file_name <- harp_get_filenames(file_date = 2017080100, template = "meps_met")
model_field <- read_netcdf(file_name, "air_temperature_2m", 0, 0)
model_field <- read_netcdf(file_name, "T2m", 0, 0)

}

