% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_netcdf_interpolate.R
\name{read_netcdf_interpolate}
\alias{read_netcdf_interpolate}
\title{Title}
\usage{
read_netcdf_interpolate(file_name, parameter = NULL, lead_time = NA_real_,
  members = NA_character_, stations = NULL)
}
\arguments{
\item{stations}{}
}
\description{
Title
}

