% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs.R
\name{read_obs}
\alias{read_obs}
\title{Read point observations from multiple files}
\usage{
read_obs(
  dttm,
  parameter,
  param_defs = getExportedValue("harpIO", "harp_params"),
  stations = NULL,
  file_path = getwd(),
  file_format = NULL,
  file_template = "vobs",
  file_format_opts = vfile_opts("vobs"),
  output_format_opts = obstable_opts(),
  return_data = FALSE,
  start_date = NULL,
  end_date = NULL,
  by = "1h",
  reads_per_write = 24,
  ...
)
}
\arguments{
\item{dttm}{A vector of date time strings to read. Can be in YYYYMMDD,
YYYYMMDDhh, YYYYMMDDhhmm, or YYYYMMDDhhmmss format. Can be numeric or
character. A vector of date-times can be generated using
\link[harpCore]{seq_dttm}.}

\item{parameter}{The names of the parameters to read. By default this is
NULL, meaning that all parameters are read from the observations files.}

\item{param_defs}{A list of parameter definitions that includes the file
format to be read. By default the built in list \code{\link{harp_params}}
is used. Modifications and additions to this list can be made using
\code{\link{modify_param_def}} and \code{\link{add_param_def}}
respectively.}

\item{stations}{The IDs of the stations to read from the files. By default
this is NULL, meaning that observations for all stations are read from the
observations files.}

\item{file_path}{The parent path to all forecast data. All file names are
generated to be under the \code{file_path} directory. The default is the
current working directory.}

\item{file_format}{The format of the files to read. By default this is
"vobs", which is the standard format used by the HIRLAM consortium. If set
to something else, \code{read_obs} will search the global environment for a
function called \code{read_<file_format>} that it will use to read from the
files.}

\item{file_template}{A template for the file names. For available built in
templates see \code{\link{show_file_templates}}. If anything else is
passed, it is returned unmodified, or with substitutions made for dynamic
values. Available substitutions are {YYYY} for year, \{MM\} for 2 digit
month with leading zero, \{M\} for month with no leading zero, and
similarly \{DD\} or \{D\} for day, \{HH\} or \{H\} for hour, \{mm\} or
\{m\} for minute. Note that the full path to the file will always be
file_path/template. Other substitutions can be passed via \code{...}}

\item{file_format_opts}{Specific options for reading the file format
specified in \code{file_format}. Should be a named list, with names
corresponding to argument for \code{read_<file_format>}.}

\item{return_data}{Logical - whether to return the data read in to the
calling environment. Due to the potential for large volumes of data, this
is set to FALSE by default.}

\item{start_date, end_date, by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The use of
\code{start_date}, \code{end_date} and \code{by} is no longer supported. \code{dttm} together
with \code{\link[harpCore]{seq_dttm}} should be used to generate equally
spaced date-times.}

\item{reads_per_write}{The number of files to read before writing out the
data to new files. Set this to a low number to reduce memory usage. The
default is 24 based on the assumption that there are observations files
every hour and writing should be done once per observation day. For the
default setting of writing to "obstable" files, this number has no impact
on the output since these files can be appended to. For other formats, this
setting might be important to prevent data from being overwritten.}

\item{...}{Other arguments to \code{generate_filenames} for getting the names
of files to read.}

\item{output_file_opts}{Options for output files. \code{read_obs} can output
data as SQLite (obstable) files or parquet (obsparquet) datasets. Use one
of the options functions to select the output file format -
\code{\link{obstable_opts()}} for SQLite fctable files (the default), or
\code{\link{obsparquet_opts()}} for a dataset of parquet files.}
}
\value{
If \code{return_data = TRUE}, a list with data frames of
observations.
}
\description{
\code{read_obs} generates file names, based on the arguments given and reads
point observations data from them. The data can optionally be re-written to
files of a different format. Due to the large volumes of data that may be
read, the function will only return data to the calling environment if
\code{return_data = TRUE}
}
\details{
\code{read_obs} is not intended to be used for reading gridded observations.
For this use \link{read_analysis} instead.
}
