% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs_convert.R
\name{read_obs_convert}
\alias{read_obs_convert}
\title{Read observations and output to sqlite OBSTABLE file(s)}
\usage{
read_obs_convert(start_date, end_date, by = "3h", obs_path = ".",
  obs_format = "vobs", obsfile_template = "vobs", parameter = NULL,
  sqlite_path = NULL,
  sqlite_template = "{sqlite_path}/OBSTABLE_{YYYY}.sqlite",
  return_data = FALSE, iterations_per_write = 24, ...)
}
\arguments{
\item{start_date}{First to read observations for. Must be at least YYYYMMDD
format.}

\item{end_date}{Last to read observations for. Must be at least YYYYMMDD
format.}

\item{by}{How frequently to read observations. Must be a chracter string with
a number followed by a letter, where "m" is minutes, "h" is hours, and "d"
is days.}

\item{obs_path}{The path for the input observation files.}

\item{obs_format}{The format of the input observation files. Currently only
"vobs".}

\item{obsfile_template}{The template for the observation files. Currently
only "vobs".}

\item{parameter}{Not used for vobs.}

\item{sqlite_path}{The path to write the sqlite output to.}

\item{sqlite_template}{The template for the sqlite observations file. The
default is "{sqlite_path}/OBSTABLE_{YYYY}.sqlite".}

\item{return_data}{Whether to return the data to the R session. The default
is FALSE.}

\item{iterations_per_write}{The number of iterations of "by" before each
write to the sqlite file. The default is 24.}

\item{...}{Arguments to read functions.}
}
\value{
If return_data is TRUE - a list with two data frames - one for synop
  (near surface) observations and one for temp (upper air) observatoions.
}
\description{
Read observations from any source (currently only vobs...) and output the
data to sqlite files.
}

