% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs_file.R
\name{read_obs_file}
\alias{read_obs_file}
\title{Read an observations file}
\usage{
read_obs_file(
  file_name,
  parameter,
  param_defs = get("harp_params"),
  dttm = NULL,
  stations = NULL,
  file_format = NULL,
  file_format_opts = list(),
  ...
)
}
\arguments{
\item{file_name}{The full path of the file to be read.}

\item{parameter}{Parameters to extract from the file. If missing or NULL, all
parameters are extracted from the file.}

\item{param_defs}{A list of parameter definitions that includes the file
format to be read. By default the built in list \code{\link{harp_params}}
is used. Modifications and additions to this list can be made using
\code{\link{modify_param_def}} and \code{\link{add_param_def}} respectively.}

\item{dttm}{The date times to extract from the file. If NULL, all date times
are extracted from the file. \code{\link[harpCore]{seq_dttm}} can be used to
generate a vector of equally spaced date-time strings.}

\item{stations}{Station IDs to extract from file. If NULL, all station IDs are
extracted from the file.}

\item{file_format}{The format of the file to read.}

\item{file_format_opts}{Options specific to the file format as a named list.}

\item{...}{Not used - absorbs other arguments from calling environment.}
}
\value{
A named list with up to four elements \itemize{ \item synop - a data
frame of observations at a single vertical level. \item temp - a data frame
of observations at multiple vertical levels. \item synop_params - a data
frame of the units and accumulation times of the parameters in synop. \item
temp_params - a data frame of units and accumulation times of the parameters
in temp.}
}
\description{
This function is for reading from an observations file. The data should be at
geographic points and either at a single vertical level, or multiple vertical
levels.
}
\details{
The file format will be guessed if it is not explicitly passed and options for
each file format should be passed as a list via the \code{file_format_opts}
argument.

No interpolation is done with this function. For reading and interpolating
gridded observations, use \link{read_analysis}.
}
