% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_point_forecast.R
\name{read_point_forecast}
\alias{read_point_forecast}
\title{Read point forecasts from an sqlite file}
\usage{
read_point_forecast(fcst_files, start_date = 197001010000,
  end_date = 300001010000, gather_data = TRUE)
}
\arguments{
\item{fcst_files}{A vector of filenames}

\item{start_date}{Start date to read from. Should be numeric or character
YYYYMMDD(HH)(mm)}

\item{end_date}{End date to read to. Should be numeric or character
YYYYMMDD(HH)(mm)}

\item{gather_data}{Whether to convert the data from wide form to lonf form
(TRUE / FALSE)}
}
\value{
A data frame
}
\description{
Reads point forecasts from an sqlite file. The forecast date column is
expected to be called fcdate. Default start and end dates are set a long time
in the past and future respectively so that if no dates are passed all data
should be read.
}

