% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_point_obs.R
\name{read_point_obs}
\alias{read_point_obs}
\title{Read point observations from OBSTABLE files}
\usage{
read_point_obs(start_date, end_date, parameter, obs_path = ".",
  obsfile_template = "obstable", gross_error_check = TRUE,
  min_allowed = NULL, max_allowed = NULL)
}
\arguments{
\item{start_date}{The start date of the observations to read.}

\item{end_date}{The end date of the observations to read.}

\item{parameter}{Which parameter to read. This will normally be a harp3
parameter name.}

\item{obs_path}{The path to the OBSTABLE files}

\item{obsfile_template}{The template for the OBSTABLE file name.}

\item{gross_error_check}{Logical of whether to perform a gross error check.}

\item{min_allowed}{The minimum value of observation to allow in the gross error
check. If set to NULL the default value for the parameter is used.}

\item{max_allowed}{The maximum value of observation to allow in the gross error
check. If set to NULL the default value for the parameter is used.}
}
\value{
A tibble with columns for validdate, SID and the parameter.
}
\description{
Read point observations from OBSTABLE files
}

