% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vfld.R
\name{read_vfld}
\alias{read_vfld}
\title{Read data from a vfld file}
\usage{
read_vfld(file_name)
}
\arguments{
\item{file_name}{Name of the vfld file}
}
\value{
A named list containing: \cr \code{synop}: the surface data, \cr
  \code{temp}: the upper air data.
}
\description{
\code{read_vfld} returns the content of a named vfld file as a named list of
data frames - synop for the the surface data (not necessarily from synop
stations, but just the content of the file), and temp for the upper air data.
}
\examples{
my_dir <- "/lustre/storeB/users/andrewts/HarpResults/vfld"
vfld_file <- harp_get_filenames(
  my_dir,
  start_date = 2017052600,
  end_date = 2017052600,
  experiment = "MEPS_summer2017_sfcPertRef",
  lead_time = 20,
  member = 0,
  file_template = "vfld")
vfld_data <- read_vfld(vfld_file)

}

