% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_point_verif.R
\name{save_point_verif}
\alias{save_point_verif}
\title{Save verification data}
\usage{
save_point_verif(
  verif_data,
  verif_path = "./Verification",
  verif_file_template =
    "{verif_path}/harpPointVerif.harp.{parameter}.harp.{start_date}-{end_date}.harp.{models}.rds",
  dir_mode = "0750",
  ...
)
}
\arguments{
\item{verif_data}{A verification object produced by one of the harpPoint
verification functions.}

\item{verif_path}{The path to save the data to. The default is a directory
called Verification inside the working directory.}

\item{verif_file_template}{A template for the verification data file. For
consistency with the default shiny app to plot verification, this should not be changed.}

\item{dir_mode}{The permissions mode to be used for creation of new
directories on Unix-alike systems. The default is "0750".}

\item{...}{Non default variables used in \code{verif_file_template}.}
}
\description{
Save verification data
}
\examples{
\dontrun{
  save_point_verif(verif_data)
}
}
