% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{seq_dates}
\alias{seq_dates}
\title{Generate a sequence of dates}
\usage{
seq_dates(start_date = NULL, end_date = NULL, by = "3h")
}
\arguments{
\item{start_date}{Start date as YYYYMMDD, YYYYMMDDhh, or YYYYMMDDhhmm. Can be
string or numeric.}

\item{end_date}{End date as YYYYMMDD, YYYYMMDDhh, or YYYYMMDDhhmm. Can be
string or numeric.}

\item{by}{Increment of the sequence. Should be a string of a number followed
by a letter, where the letter gives the units - may be d for days, h for
hours or m for minutes.}
}
\value{
A sequence of date-times in YYYYMMDDhhmm format
}
\description{
Generate a sequence of dates
}
\examples{
seq_dates()
seq_dates(20170101, 20170131, by = "1d")
seq_dates(201701010600, 201701051200)
}

