% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{show_file_templates}
\alias{show_file_templates}
\title{Show built in file templates}
\usage{
show_file_templates(template_row = NULL)
}
\arguments{
\item{template_row}{Some of the templates are too wide to show on the screen,
so to show a template in full, pass the row number of the template you want
to show in full in \code{template_row}}
}
\description{
Show built in file templates
}
\examples{
show_file_templates()
show_file_templates(1)
show_file_templates(18)
}
