% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_opts.R
\name{sqlite_opts}
\alias{sqlite_opts}
\title{Options for Writing SQLite files}
\usage{
sqlite_opts(
  path = NULL,
  template = "fctable_det",
  index_cols = c("fcdate", "leadtime", "SID"),
  synchronous = c("off", "normal", "full", "extra"),
  journal_mode = c("delete", "truncate", "persist", "memory", "wal", "off"),
  remove_model_elev = FALSE
)
}
\arguments{
\item{path}{If not NULL, sqlite files are generated and written to the
directory specified here.}

\item{template}{The template for the filenames of the sqlite files. See
\code{\link{show_file_templates}} for available built in templates - for
point forecast sqlite files, these are templates beginning "fctable_". The
default is "fctable_det".}

\item{index_cols}{The columns to index by}

\item{synchronous}{The synchronus setting for sqlite files. The defualt is
"off", but could also be "normal", "full", or "extra". See
\url{https://www.sqlite.org/pragma.html#pragma_synchronous} for more
information.}

\item{journal_mode}{The journal mode for the sqlite files. The default is
"delete", but can also be "truncate", "persist", "memory", "wal", or "off".
See \url{https://www.sqlite.org/pragma.html#pragma_journal_mode} for more
information.}

\item{remove_model_elev}{Set to TRUE to not include model elevation in the
sqlite output files. For multi model ensembles, members having different
model elevations from each other will make it impossible to include all
members in the same row and thus break unique constraints for the row
indexing.}
}
\value{
A list with options for writing sqlite files
}
\description{
Options for Writing SQLite files
}
\examples{
sqlite_opts()
sqlite_opts(path = tempdir())
sqlite_opts(path = tempdir(), template = "fctable_eps_all_leads")
}
