% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_groups}
\alias{station_groups}
\title{Weather station IDs with regional groups}
\format{
A data frame with 6001 rows and 2 variables: \describe{
\item{SID}{station ID number} \item{group}{Geographic group the station
belongs to}}
}
\source{
selecttion.pm from HARMONIE / HIRLAM monitor
}
\usage{
station_groups
}
\description{
A dataset with station IDs and geographic groups they belong it. Can be used
to do grouped verification by joining the station_groups to a harp_fcst list
and running the verification with \code{groupings = c("leadtime", "group")}.
Note that many stations belong to more than one group
}
\keyword{datasets}
