% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_list}
\alias{station_list}
\title{List of Weather Stations}
\format{
A data frame with 13417 rows and 5 variables: \describe{
\item{SID}{station ID number} \item{lat}{latitude of the station, in
decimal degrees} \item{lon}{longitude of the station, in decimal degrees}
\item{elev}{elevation of the station, in metres} \item{name}{the name of
the station}}
}
\source{
HIRLAM station list
}
\usage{
station_list
}
\description{
A dataset containing site ID, latitude, longitude, elevation and name of
13417 weather stations world wide.
}
\keyword{datasets}
