% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{str_datetime_to_unixtime}
\alias{str_datetime_to_unixtime}
\title{Convert a date-time string to unix time}
\usage{
str_datetime_to_unixtime(str_datetime)
}
\arguments{
\item{str_datetime}{A date-time string - its format is guessed from the
number of characters in the string: 8 = YYYYMMDD, 10 = YYYYMMDDHH, 12 =
YYYYMMDDHHmm. May be passed as a character string or numeric.}
}
\value{
The unix time in seconds since 01/01/1970 00:00:00
}
\description{
This is a convenient wrapper for the YMD..2unix functions
}
\examples{
str_datetime_to_unixtime("20170101")
str_datetime_to_unixtime("2017010112")
str_datetime_to_unixtime("201701011230")

}

