% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_from_q_and_p.R
\name{td_from_q_and_p}
\alias{td_from_q_and_p}
\title{Compute dewpoint temperature from specific humidity and pressure}
\usage{
td_from_q_and_p(q, p, kelvin = TRUE)
}
\arguments{
\item{q}{Specific humidity - can be in g / kg or kg / kg}

\item{p}{pressure - can be Pa or hPa (mb)}

\item{kelvin}{Logical - if TRUE (the defualt) return temperature in Kelvin
otherwise in degrees C.}
}
\value{
Dew point temperature
}
\description{
Compute dewpoint temperature from specific humidity and pressure
}
\examples{
td_from_q_and_p(0.008, 1000)
td_from_q_and_p(8, 10000, kelvin = FALSE)
}
