% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{unix2datetime}
\alias{unix2datetime}
\title{Convert unix time to date type}
\usage{
unix2datetime(x)
}
\arguments{
\item{x}{unix time in seconds since 01/01/1970 00:00:00}
}
\value{
A date type object
}
\description{
Convert unix time to date type
}
\examples{
unix2datetime(as.numeric(Sys.time()))
}

