% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{unix2YMD}
\alias{unix2YMD}
\title{Convert unix time to YYYYMMDD string}
\usage{
unix2YMD(x)
}
\arguments{
\item{x}{unix time in seconds since 01/01/1970 00:00:00}
}
\value{
A YYYYMMDD string
}
\description{
Convert unix time to YYYYMMDD string
}
\examples{
unix2YMD(as.numeric(Sys.time()))
}
