% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{unix2YMDh}
\alias{unix2YMDh}
\title{Convert unix time to YYYYMMDDHH string}
\usage{
unix2YMDh(x)
}
\arguments{
\item{x}{unix time in seconds since 01/01/1970 00:00:00}
}
\value{
A YYYYMMDDHH string
}
\description{
Convert unix time to YYYYMMDDHH string
}
\examples{
unix2YMDh(as.numeric(Sys.time()))
}

