% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{unix2YMDhm}
\alias{unix2YMDhm}
\title{Convert unix time to YYYYMMDDHHmm string}
\usage{
unix2YMDhm(x)
}
\arguments{
\item{x}{unix time in seconds since 01/01/1970 00:00:00}
}
\value{
A YYYYMMDDHHmm string
}
\description{
Convert unix time to YYYYMMDDHHmm string
}
\examples{
unix2YMDhm(as.numeric(Sys.time()))
}

