% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{unixtime_to_str_datetime}
\alias{unixtime_to_str_datetime}
\title{Convert unix time to a date-time string}
\usage{
unixtime_to_str_datetime(unixtime, str_function)
}
\arguments{
\item{unixtime}{The unix time: number of seconds since 01/01/1970 00:00:00
UTC}

\item{str_function}{The date-time string format to convert to. Available
options are YMD, YMDh and YMDhm where Y is year (4 digits), M is month (2
digits), D is day (2 digits), h is hour (2 digits) and m is minutes (2
digits).}
}
\value{
A date-time string
}
\description{
This is a convenient wrapper for the unix2YMD... functions
}
\examples{
unixtime_to_str_datetime(as.numeric(Sys.time()), YMD)
unixtime_to_str_datetime(as.numeric(Sys.time()), YMDh)
unixtime_to_str_datetime(as.numeric(Sys.time()), YMDhm)

}
