% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_v_parameter_synop.R,
%   R/parse_v_parameter_temp.R
\name{v_parameters}
\alias{v_parameters}
\alias{parse_v_parameter_synop}
\alias{parse_v_parameter_temp}
\title{Parameters for vobs and vfld files}
\usage{
parse_v_parameter_synop(v_parameter)

parse_v_parameter_temp(v_parameter)
}
\arguments{
\item{v_parameter}{vobs or vfld parameter}
}
\value{
A list with the harp paramater name and the units of the parameter.
}
\description{
Convert parameter names from vobs or vfld files into harp formate parameters.
Use parse_v_parameter_synop for SYNOP parameters and parse_v_parameter_temp
for TEMP parameters.
}
\examples{
parse_v_parameter_synop("TT")
parse_v_parameter_synop("RH")
parse_v_parameter_synop("FI")
parse_v_parameter_synop("PE")

parse_v_parameter_temp("TT")
parse_v_parameter_temp("RH")
parse_v_parameter_temp("FI")
parse_v_parameter_temp("PP")

}
