% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fctable.R
\name{write_fctable}
\alias{write_fctable}
\title{Title}
\usage{
write_fctable(tableData, fctable_path, fcstExp, parameter,
  filename_template = "${fctable_path}/${fcstExp}/${YYYY}/${MM}/FCTABLE_${parameter}_${YYYY}${MM}_${HH}+${LDT3}.sqlite")
}
\arguments{
\item{filename_template}{}
}
\description{
Title
}

