% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fctable_to_sqlite.R
\name{write_fctable_to_sqlite}
\alias{write_fctable_to_sqlite}
\title{Title}
\usage{
write_fctable_to_sqlite(data, filename, tablename = "FC",
  primary_key = "SID, fcdate, leadtime")
}
\arguments{
\item{primary_key}{}
}
\description{
Title
}

