% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{YMD}
\alias{YMD}
\title{Convert date type to YYYYMMDD string}
\usage{
YMD(x)
}
\arguments{
\item{x}{A date object.}
}
\value{
A YYYYMMDD string
}
\description{
Convert date type to YYYYMMDD string
}
\examples{
YMD(Sys.time())
}
