% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{YMD2unix}
\alias{YMD2unix}
\title{Convert YYYYMMDD string to unix time}
\usage{
YMD2unix(x)
}
\arguments{
\item{x}{A YYYYMMDD string}
}
\value{
unix time in seconds since 01/01/1970 00:00:00
}
\description{
Convert YYYYMMDD string to unix time
}
\examples{
YMD2unix(20170101)
}
