% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{YMDh}
\alias{YMDh}
\title{Convert date type to YYYYMMDDHH string}
\usage{
YMDh(x)
}
\arguments{
\item{x}{A date object.}
}
\value{
A YYYYMMDDHH string
}
\description{
Convert date type to YYYYMMDDHH string
}
\examples{
YMDh(Sys.time())
}

