% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{YMDh2unix}
\alias{YMDh2unix}
\title{Convert YYYYMMDDHH string to unix time}
\usage{
YMDh2unix(x)
}
\arguments{
\item{x}{A YYYYMMDDHH string}
}
\value{
unix time in seconds since 01/01/1970 00:00:00
}
\description{
Convert YYYYMMDDHH string to unix time
}
\examples{
YMDh2unix(2017010112)
}

