% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{YMDhm}
\alias{YMDhm}
\title{Convert date type to YYYYMMDDHHmm string}
\usage{
YMDhm(x)
}
\arguments{
\item{x}{A date object.}
}
\value{
A YYYYMMDDHHmm string
}
\description{
Convert date type to YYYYMMDDHHmm string
}
\examples{
YMDhm(Sys.time())
}
