% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{YMDhm2unix}
\alias{YMDhm2unix}
\title{Convert YYYYMMDDHHmm string to unix time}
\usage{
YMDhm2unix(x)
}
\arguments{
\item{x}{A YYYYMMDDHHmm string}
}
\value{
unix time in seconds since 01/01/1970 00:00:00
}
\description{
Convert YYYYMMDDHHmm string to unix time
}
\examples{
YMDhm2unix(201701011230)
}
