% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_dates.R
\name{YMDhms}
\alias{YMDhms}
\title{Convert date type to YYYYMMDDHHmm string}
\usage{
YMDhms(x)
}
\arguments{
\item{x}{A date object.}
}
\value{
A YYYYMMDDHHmm string
}
\description{
Convert date type to YYYYMMDDHHmm string
}
\examples{
YMDhms(Sys.time())
}
