% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grib_param_info.R
\name{get_grib_param_info}
\alias{get_grib_param_info}
\title{Internal function to get grib parameter information}
\usage{
get_grib_param_info(param, vertical_coordinate = NA_character_)
}
\arguments{
\item{param}{Parameter name (or harp_param object)}

\item{vertical_coordinate}{The vertical coordinate for upper air data. May be
"pressure", "model" or "height".}
}
\value{
A list with \code{short_name}, \code{param_number},
\code{level_type}, \code{level_numbe.r}
}
\description{
Gets the shortname and parameter number for grib files. For
surface and screen level parameters the level type and level number are
assumed. For upper air parameters the level and, optionally, the level type
must be supplied. The default is for pressure levels.
}
