% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grib_param_info.R
\name{get_grib_param_info}
\alias{get_grib_param_info}
\title{Get grib paramater information}
\usage{
get_grib_param_info(param, level = NULL, levtype = 100)
}
\arguments{
\item{param}{Parameter name}

\item{level}{The level of the field required - may be pressure level, hybrid
level or height above ground.}

\item{levtype}{The level type given as a grib \code{indicatorOfTypeOfLevel}
integer: 100 = pressure levels, 105 = height above ground, 109 = hybrid
levels.}
}
\value{
A lisy with \code{short_name}, \code{param_number},
  \code{level_type}, \code{level_numbe.r}
}
\description{
Gets the shortname and parameter number for a HarmonEPS grib file. For
surface and screen level parameters the level type and level number are
assumed. For upper air parameters the level and, optionally, the level type
must be supplied. The default is for pressure levels.
}
\examples{
get_grib_param_info("pcp")
get_grib_param_info("t2m")
get_grib_param_info("t", level = 500)
get_grib_param_info("rh", level = 32, levtype = 109)
}

