% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netcdf_param_info.R
\name{get_netcdf_param_info}
\alias{get_netcdf_param_info}
\title{Get MET Norway NetCDF parameter name from a HARP parameter name}
\usage{
get_netcdf_param_info(param, vc = NA_character_, opts = netcdf_opts())
}
\arguments{
\item{param}{HARP parameter name.}

\item{vc}{The vertical coordinate. Set to NA for no vertical coordinate,
otherwise "pressure" or "model".}
}
\value{
The parameter name in a MET Norway NetCDF file
}
\description{
Get MET Norway NetCDF parameter name from a HARP parameter name
}
\examples{
get_netcdf_param_MET("T2m")
get_netcdf_param_MET("PMSL")
get_netcdf_param_MET("S10m")
get_netcdf_param_MET("Pcp")
}
