% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netcdf_param_info.R
\name{get_netcdf_param_info}
\alias{get_netcdf_param_info}
\title{Internal function to get NetCDF parameter name from a HARP parameter name}
\usage{
get_netcdf_param_info(
  param,
  vc = NA_character_,
  opts = netcdf_opts(),
  param_defs = get("harp_params")
)
}
\arguments{
\item{param}{HARP parameter name.}

\item{vc}{The vertical coordinate. Set to NA for no vertical coordinate,
otherwise "pressure" or "model".}

\item{opts}{Options for netcdf files as generated by
\code{\link{netcdf_opts}}. When certain options_set values are passed the
options may be modified depending on the parameter.}
}
\value{
A list with the harp parameter name, the netcdf parameter name and
the netcdf options, that may have been modified.
}
\description{
Internal function to get NetCDF parameter name from a HARP parameter name
}
