% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_opts.R
\name{grib_opts}
\alias{grib_opts}
\title{Set options for grib files}
\usage{
grib_opts(meta = TRUE, multi = FALSE)
}
\arguments{
\item{meta}{Logical. Weather to read metadata}

\item{multi}{Logical. Whether there are multi field grib messages.}
}
\value{
A list
}
\description{
Set options for grib files
}
\examples{
grib_opts()
}
