% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msub.R
\name{msub}
\alias{msub}
\title{Multiple pattern matching and replacement}
\usage{
msub(x, pattern, replacement, ...)
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A character vector of patterns to replace.}

\item{replacement}{A character vector of the same length as pattern with the
corresponding replacements.}

\item{...}{Other arguments to \link[base]{gsub}}
}
\value{
A character vector of the same length as x
}
\description{
\code{msub} calls \link[base]{gsub} for each pattern / replacement pair.
Where a pattern is not found, nothing is replaced for that pattern /
replacement pair.
}
\examples{
msub(letters[1:10], c("a", "c", "e"), c("A", "C", "E"))

}
