% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_harp_parameter.R
\name{parse_harp_parameter}
\alias{parse_harp_parameter}
\title{Parse the parameter name and decide on level, accumulation...}
\usage{
parse_harp_parameter(param)
}
\arguments{
\item{param}{Parameter name}
}
\value{
A list with \code{fullname}, \code{basename}, \code{level_type},
  \code{level}, \code{accum}
}
\description{
Parses a harp-style parameter name and returns accumulation time, level (if
any) and the base name of the parameter
}
\examples{
parse_harp_param("AccPcp1h")
parse_harp_param("z500")
parse_harp_param("s100m")
}

