% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fa.R
\name{read_fa}
\alias{read_fa}
\title{Read a field from an FA file}
\usage{
read_fa(file_name, parameter, meta = TRUE, fa_type = "arome",
  fa_vector = TRUE, ...)
}
\arguments{
\item{file_name}{The FA file name. "file@arch" signifies a file inside a tar archive.
It may also be a \code{FAfile} object.}

\item{parameter}{The parameter to read. Standard HARP names are used, but full FA field names will also
work.}

\item{meta}{If TRUE, also read all meta data (domain, time properties).}

\item{fa_type}{The kind of model file: "arome", "alaro", "surfex"...}

\item{fa_vector}{TRUE if the wind variable (speed, direction) must be calculated from components}

\item{...}{Ignored}
}
\value{
A 2d geofield object (2d array with projection information)
}
\description{
Read a field from an FA file
}
\examples{
model_geofield <- read_fa(file_name, "t2m")
model_geofield <- read_fa(file_name, "t500")
model_geofield <- read_fa(file_name, "topo")
}
