% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fa.R
\name{read_fa}
\alias{read_fa}
\title{Read a field from a FA file}
\usage{
read_fa(infile, parameter, meta = TRUE, fa_type = "alaro",
  fa_vector = FALSE, ...)
}
\arguments{
\item{parameter}{The parameter to read. Standard HARP names are used.}

\item{meta}{If TRUE, also read all meta data (domain, time properties).}

\item{fa_type}{The kind of model file: "arome", "alaro", "surfex"...}

\item{fa_vector}{TRUE if the wind variable (speed, direction) must be calculated from components}

\item{...}{Arguments for \code{Rfa::FAdec}}

\item{filename}{The FA file name.}
}
\value{
A geofield object with 2d array and projection information
}
\description{
Read a field from a FA file
}
\examples{
model_geofield <- read_fa(file_name, "t2m")
model_geofield <- read_fa(file_name, "t500")
model_geofield <- read_fa(file_name, "topo")
}

