% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fa.R
\name{read_fa_interpolate}
\alias{read_fa_interpolate}
\title{Read FA files & interpolate}
\usage{
read_fa_interpolate(file_name, parameter, lead_time = 0, member = NULL,
  model = NULL, stations = NULL, method = "closest",
  use_mask = FALSE, ...)
}
\arguments{
\item{parameter}{The parameter(s) to be decoded.}

\item{lead_time}{The lead time(s) to be extracted. May be a vector!}

\item{member}{Not used. Only there for API reasons.}

\item{model}{Name of the model. Used to find domain and interpolation weights.}

\item{...}{Extra arguments for read_fa[tar]}

\item{filename}{Name of a tar archive containing FA files}

\item{iweights}{Interpolation weights (and domain information) can also be passed explicitely.}

\item{lead_time}{Not used, only added to to the output table. Must be a single number.}
}
\value{
A tibble with interpolated forecasts for the stations list
}
\description{
Read FA files & interpolate
}
