% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grib.R
\name{read_grib}
\alias{read_grib}
\title{Read a field from a grib file}
\usage{
read_grib(filename, parameter, ...)
}
\arguments{
\item{filename}{The grib file name.}

\item{parameter}{The parameter to read. Standard HARP names are used.}

\item{...}{Arguments for \code{get_grib_param_info} for level and levtype.}
}
\value{
A geofield object with 2d array and projection information
}
\description{
Read a field from a grib file
}
\examples{
file_name <- "/lustre/storeB/users/andrewts/mepsr_data/grib/fc2017052600+001grib_fp"
model_geofield <- read_grib(file_name, "t2m")
model_geofield <- read_grib(file_name, "t", level = 500)
model_geofield <- read_grib(file_name, "topo")
}

