% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grib.R
\name{read_grib_interpolate}
\alias{read_grib_interpolate}
\title{Read a field from a grib file & interpolate}
\usage{
read_grib_interpolate(file_name, parameter, lead_time, members = NULL,
  init = list(), method = "closest", use_lsm = FALSE, ...)
}
\arguments{
\item{parameter}{The parameter to read. Standard HARP names are used.}

\item{lead_time}{lead time}

\item{members}{ens members}

\item{init}{Initialisation for interpolation. A list that contains
station locations and (possibly) pre-calculated interpolation weights etc.}

\item{...}{Arguments for \code{Rgrib2::Gdec}}

\item{filename}{The grib file name.}

\item{meta}{If TRUE, also read all meta data (domain, time properties).}
}
\value{
A tibble
}
\description{
Read a field from a grib file & interpolate
}
