% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grid.R
\name{read_grid}
\alias{read_grid}
\title{Read gridded data in various formats.}
\usage{
read_grid(filename, file_format = NULL, ...)
}
\arguments{
\item{filename}{A character string with the (full) file name.}

\item{...}{All arguments passed to the specified reader function.}

\item{format}{The file format. Possible values include grib, netcdf, FA, 
hdf5... Whatever the value is, it is suposed to correspond to a function
"read_XXX" that can deal wit hthe format.}
}
\value{
A geofield or (possibly) a plain  matrix.
}
\description{
Read gridded data in various formats.
}
