% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs.R
\name{read_obs}
\alias{read_obs}
\title{Title}
\usage{
read_obs(obs_files, start_date, end_date, parameter, gross_error_check = TRUE,
  min_allowed = NULL, max_allowed = NULL)
}
\arguments{
\item{obs_files}{A vector of observation files in sqlite format to read.}

\item{start_date}{The start date of the observations to read.}

\item{end_date}{The end date of the observations to read.}

\item{parameter}{Which parameter to read. This will normally be a harp3
parameter name.}

\item{gross_error_check}{Logical of whether to perform a gross error check.}

\item{min_allowed}{The minimum value of observation to allow in the gross error
check. If set to NULL the default value for the parameter is used.}

\item{max_allowed}{The maximum value of observation to allow in the gross error
check. If set to NULL the default value for the parameter is used.}
}
\value{
A tibble with columns for validdate, SID and the parameter.
}
\description{
Title
}

