% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_point_obs.R
\name{read_point_obs}
\alias{read_point_obs}
\title{Read point observations from OBSTABLE files}
\usage{
read_point_obs(
  dttm,
  parameter,
  file_path = ".",
  file_template = "obstable",
  file_format = "obstable",
  gross_error_check = TRUE,
  min_allowed = NULL,
  max_allowed = NULL,
  stations = NULL,
  vertical_coordinate = c(NA_character_, "pressure", "model", "height"),
  start_date = NULL,
  end_date = NULL,
  by = "1h",
  obs_path = NULL,
  obsfile_template = NULL
)
}
\arguments{
\item{dttm}{A vector of date time strings to read. Can be in YYYYMMDD,
YYYYMMDDhh, YYYYMMDDhhmm, or YYYYMMDDhhmmss format. Can be numeric or
character. \code{\link[harpCore]{seq_dttm}} can be used to generate a
vector of equally spaced date-time strings. Alternatively valid date-times
can be extracted from a forecast data frame or \code{harp_list} of forecast
data frames using \code{\link[harpCore]{unique_valid_dttm}}.}

\item{parameter}{Which parameter to read. This will normally be a harp
parameter name.}

\item{file_path}{The path to the observation files / dataset}

\item{file_template}{The template for the observations file name.}

\item{file_format}{The format of the files storing the observations. Can be
"obstable" for \emph{SQLite} files with the same schema as that produced by
\code{read_obs()} with \code{file_format_opts = obstable_opts()}, or "obsparquet" for
an \emph{Arrow} dataset of _parquet_files with the same schema as that produced
by \code{read_obs()} with \code{file_format_opts = obsparquet_opts()}.}

\item{gross_error_check}{Logical of whether to perform a gross error check.}

\item{min_allowed}{The minimum value of observation to allow in the gross
error check. If set to NULL the default value for the parameter is used.}

\item{max_allowed}{The maximum value of observation to allow in the gross
error check. If set to NULL the default value for the parameter is used.}

\item{stations}{The stations to retrieve observations for. This should be a
vector of station ID numbers. Set to NULL to retrieve all stations.}

\item{vertical_coordinate}{If upper air for multiple levels are to be read,
the vertical coordinate of the data is given here. The default is
"pressure", but can also be "model" for model levels, or "height" for
height above ground /sea level.}

\item{start_date, end_date, by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The use of
\code{start_date}, \code{end_date} and \code{by} is no longer supported. \code{dttm} together
with \code{\link[harpCore]{seq_dttm}} should be used to generate equally
spaced date-times.}

\item{obs_path, obsfile_template}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} These
arguments have been renamed \code{file_path} and \code{file_template} for consistency
with other harpIO functions. Please use the new argument names.}
}
\value{
A tibble with columns for valid_dttm, SID and the parameter.
}
\description{
Data are read directly from OBSTABLE files with no modification except in the
case of precipitation accumulations. Where possible, observations that are
not available for the given accumulation time, they are derived from
observations for other accumulation times, either by subtraction or addition.
}
\examples{
if (requireNamespace("harpData", quietly = TRUE)) {
  read_point_obs(
    seq_dttm(2019021700, 2019022023),
    "T2m",
    obs_path = system.file("OBSTABLE", package = "harpData")
  )

  # stations can be specified using a vector of station ID numbers
  read_point_obs(
    seq_dttm(2019021700, 2019022023),
    "T2m",
    obs_path = system.file("OBSTABLE", package = "harpData"),
    stations = c(1001, 1010)
  )

  # Gross error checks are done automatically but the allowable values
  # can be changed with min_allowed and max_allowed.
  obs <- read_point_obs(
    seq_dttm(2019021700, 2019022023),
    "T2m",
    obs_path = system.file("OBSTABLE", package = "harpData"),
    min_allowed = 260,
    max_allowed = 280
  )

  # The removed observations are stored in the attribute "bad_obs"
  attr(obs, "bad_obs")

  # For vertical profiles, the vertical coordinate must be specified
  read_point_obs(
    seq_dttm(2019021700, 2019022023),
    "Z",
    obs_path            = system.file("OBSTABLE", package = "harpData"),
    vertical_coordinate = "pressure"
  )
}

}
