% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vobs.R
\name{read_vobs}
\alias{read_vobs}
\title{Read data from a vobs file.}
\usage{
read_vobs(file_name, missing_value = -99, ...)
}
\arguments{
\item{file_name}{Name of the vobs file.}

\item{missing_value}{Missing value indicator in vobs file.}

\item{...}{Not used. Absorbs unused arguments.}
}
\value{
A list with data frames for synop and temp data.
}
\description{
\code{read_vobs} returns the content of a named vobs file as a list of
data frames - one for synop data and one for temp (upper air) data.
}
\examples{

}
