% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fa.R
\name{rh2tdew}
\alias{rh2tdew}
\title{Standard Magnus formula for dewpoint temperature}
\usage{
rh2tdew(T, RH, tuning = c(a = 17.67, b = 243.5))
}
\arguments{
\item{T}{Temperature in degrees Celsius (Kelvin is also OK). Could be a geofield.}

\item{RH}{Relative humidity 0..1 (percentage is also OK).}

\item{tuning}{A vector with two tuning parameters for the formula. Default is from NOAA.}
}
\value{
Dewpoint temperature
}
\description{
Standard Magnus formula for dewpoint temperature
}

