% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fctable_to_sqlite.R
\name{write_fctable_to_sqlite}
\alias{write_fctable_to_sqlite}
\title{Write a \code{harp_fcst} object to an sqlite file.}
\usage{
write_fctable_to_sqlite(data, filename, tablename = "FC",
  primary_key = "SID, fcdate, leadtime")
}
\arguments{
\item{data}{The fcst data to write.}

\item{filename}{The name of the sqlite file to write the data to.}

\item{tablename}{The table in the sqlite file to write the data to.}

\item{primary_key}{The primary key to be used.}
}
\value{
This function only has a side effect (writing to sqlite file).
}
\description{
This function should rarely be used as a standalone function. It is called by
\code{read_det_interpolate} and \code{read_eps_interpolate}.
}

